requirements.txt
Django==5.2.8
djangorestframework
djangorestframework-simplejwt
social-auth-app-django
django-cors-headers
